Lots of improvements
